Relocate InviteWorkerToCompany presenter + cntrllr
