+++ arbeitszeit_web/controllers/invite_worker_to_company_controller.py
@@ -1,11 +1,9 @@
 from dataclasses import dataclass
-from typing import List, Protocol
+from typing import Protocol
 from uuid import UUID
 
 from arbeitszeit.use_cases.invite_worker_to_company import InviteWorkerToCompanyUseCase
-from arbeitszeit_web.translator import Translator
-
-from .session import Session
+from arbeitszeit_web.session import Session
 
 
 class InviteWorkerToCompanyForm(Protocol):
@@ -36,26 +34,3 @@ class InviteWorkerToCompanyController:
             return UUID(form.get_worker_id())
         except ValueError:
             raise ValueError("worker_id is not a valid UUID")
-
-
-@dataclass
-class ViewModel:
-    notifications: List[str]
-
-
-@dataclass
-class InviteWorkerToCompanyPresenter:
-    translator: Translator
-
-    def present(
-        self, use_case_response: InviteWorkerToCompanyUseCase.Response
-    ) -> ViewModel:
-        if use_case_response.is_success:
-            notifications = [
-                self.translator.gettext("Worker has been invited successfully.")
-            ]
-        else:
-            notifications = [self.translator.gettext("Worker could not be invited.")]
-        return ViewModel(
-            notifications=notifications,
-        )
